requirements.txt
discord.py[voice]
ffmpeg
asyncpraw
youtube_dl